+++ com.unity.render-pipelines.high-definition/Runtime/Debug/DebugDisplay.cs
@@ -219,6 +219,11 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         {
             return data.mipMapDebugSettings.IsDebugDisplayEnabled();
         }
+        public bool IsMatcapViewEnabled(HDCamera camera)
+        {
+            bool sceneViewLightingDisabled = CoreUtils.IsSceneLightingDisabled(camera.camera);
+            return sceneViewLightingDisabled || GetDebugLightingMode() == DebugLightingMode.MatcapView;
+        }
 
         private void DisableNonMaterialDebugSettings()
         {